+++ packages/dash-core-components/dash_core_components/Graph.py
@@ -11,18 +11,25 @@ Keyword arguments:
 - id (string; optional): The ID of this component, used to identify dash components
 in callbacks. The ID needs to be unique across all of the
 components in an app.
-- clickData (dict; optional): Data from latest click event
-- clickAnnotationData (dict; optional): Data from latest click annotation event
-- hoverData (dict; optional): Data from latest hover event
+- clickData (dict; optional): Data from latest click event. Read-only.
+- clickAnnotationData (dict; optional): Data from latest click annotation event. Read-only.
+- hoverData (dict; optional): Data from latest hover event. Read-only.
 - clear_on_unhover (boolean; optional): If True, `clear_on_unhover` will clear the `hoverData` property
 when the user "unhovers" from a point.
 If False, then the `hoverData` property will be equal to the
 data from the last point that was hovered over.
-- selectedData (dict; optional): Data from latest select event
+- selectedData (dict; optional): Data from latest select event. Read-only.
 - relayoutData (dict; optional): Data from latest relayout event which occurs
-when the user zooms or pans on the plot
-- restyleData (dict; optional): Data from latest restyle event which occurs
-when the user toggles a legend item
+when the user zooms or pans on the plot or other
+layout-level edits. Has the form `{<attr string>: <value>}`
+describing the changes made. Read-only.
+- restyleData (list; optional): Data from latest restyle event which occurs
+when the user toggles a legend item, changes
+parcoords selections, or other trace-level edits.
+Has the form `[edits, indices]`, where `edits` is an object
+`{<attr string>: <value>}` describing the changes made,
+and `indices` is an array of trace indices that were edited.
+Read-only.
 - figure (dict; optional): Plotly `figure` object. See schema:
 https://plot.ly/javascript/reference
 Only supports `data` array and `layout` object.